Issue#11: Render proper view on user login based on their permissions.
The navbar and footer will be loaded by Router::getLayoutContent() based on the $layout suffix.

Router::getComponentContent() instead of requiring the component it will load. If the file doesn't exist the warning will be suppressed. This will be used with fallback on Router::getLayoutContent().

Seperated Home view with home.auth which will be rendered only when user is login.

The layout of SiteController will change from main to admin if user is login.

For now, when the user is login the rendered view will be rendered by default.